requirements.txt
click==7.1.1
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1
pydantic==1.5.1
pymongo==3.10.1
python-dotenv==0.13.0
starlette==0.13.2
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
